+++ internal/provider/provider_next.go
@@ -164,6 +164,7 @@ func (p *frameworkProvider) Resources(ctx context.Context) []func() resource.Res
 		NewTeamNotificationConfigurationResource,
 		NewTestVariableResource,
 		NewWorkspaceRunTaskResource,
+		NewNotificationConfigurationResource,
 	}
 }